+++ src/main/java/nl/ramsolutions/sw/magik/jacoco/generators/CoberturaXmlReportGenerator.java
@@ -31,6 +31,7 @@ public class CoberturaXmlReportGenerator extends BaseReportGenerator {
      * @param executionDataFile File to {@literal jacoco.exec}.
      * @param outputFile File to report file (XML).
      * @param discardExecutable Discard executable.
+     * @param discardNonMagik Discard non-Magik code.
      * @param bundleName Name of the bundle.
      */
     public CoberturaXmlReportGenerator(
@@ -39,8 +40,9 @@ public class CoberturaXmlReportGenerator extends BaseReportGenerator {
             final File executionDataFile,
             final File outputFile,
             final boolean discardExecutable,
+            final boolean discardNonMagik,
             final String bundleName) {
-        super(productPaths, sourcePaths, executionDataFile, outputFile, discardExecutable, bundleName);
+        super(productPaths, sourcePaths, executionDataFile, outputFile, discardExecutable, discardNonMagik, bundleName);
     }
 
     @Override